updated AI.cs
no goes to the last seen position of object where it was last seen
