+++ Assets/Scenes/Pathfinding.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8e68b351c7734e04eaf5624426227cc2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
