+++ Assets/Scenes/Pathfinding/NavMesh.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7eb738a19988aaa45ae7f0538c34498c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
