+++ Assets/Scripts/AI/moving kek.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 289361a071f98b74ca1f8221887f08f3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
