+++ Assets/Scripts/AI/test.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b1f2bfe3f9616f34d8d16b14ff21005a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
